Версия готова к использованию. Отсутствует возможность установки порогов через ModbusTCP сервер

+++ Arduino/ThermalChamber/Config.h
@@ -4,13 +4,13 @@
 
 #define SERIAL_DEBUG_ENABLED
 
-#define LED_PIN           2
-#define SELECT_MODE_PIN   4
-#define T_H_SENSOR_PIN    5/*D1*/
-#define OUT1_HIGH_PIN     12
-#define OUT1_LOW_PIN      13
-#define OUT2_HIGH_PIN     14
-#define OUT2_LOW_PIN      15
+#define LED_PIN           (int)2
+#define SELECT_MODE_PIN   (int)4
+#define T_H_SENSOR_PIN    (int)5/*D1*/
+#define OUT1_HIGH_PIN     (int)12
+#define OUT1_LOW_PIN      (int)13
+#define OUT2_HIGH_PIN     (int)14
+#define OUT2_LOW_PIN      (int)15
 
 #define A_IN              A0  /* ESP8266 Analog Pin ADC0 = A0*/
